pyproject.toml
[project]
name = "mcp-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-openai>=0.3.11",
    "langgraph>=0.3.21",
    "openai>=1.69.0",
]
